requirements.txt
# Web Framework & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Data Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database & ORM
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.1
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-jose[cryptography]>=3.3.0
PyYAML>=6.0
reportlab>=4.0.0

# Audio Processing (Gentle integration)
aiohttp>=3.9.0
aiofiles>=23.0.0

# Audio Processing (Conversation Practice)
faster-whisper>=0.10.0
edge-tts>=6.1.9

# Testing
pytest>=8.0.0
httpx>=0.26.0
pytest-asyncio>=0.23.5